Don't require `click` for Docker platform extraction
